Refactors parts of the PredictPostings decorator into a new abstract base class called SmartImporterDecorator.
