+++ smart_importer/predict_postings.py
@@ -7,7 +7,7 @@ import logging
 from functools import wraps
 from typing import List, Union
 
-from beancount.core.data import Transaction, TxnPosting, filter_txns
+from beancount.core.data import Transaction, filter_txns
 from beancount.ingest.cache import _FileMemo
 from sklearn.feature_extraction.text import CountVectorizer
 from sklearn.pipeline import Pipeline, FeatureUnion
@@ -16,13 +16,14 @@ from sklearn.svm import SVC
 from smart_importer import machinelearning_helpers as ml
 
 # configure logging
+from smart_importer.decorator_baseclass import SmartImporterDecorator
+
 LOG_LEVEL = logging.DEBUG
 logging.basicConfig(level=LOG_LEVEL)
 logger = logging.getLogger(__name__)
 
 
-# the decorator
-class PredictPostings:
+class PredictPostings(SmartImporterDecorator):
     '''
     Applying this decorator to a beancount importer or its extract method
     will predict and auto-complete missing second postings
@@ -57,53 +58,9 @@ class PredictPostings:
         self.predict_second_posting = predict_second_posting
         self.suggest_accounts = suggest_accounts
 
-    def __call__(self, to_be_decorated=None, *args, **kwargs):
 
-        if inspect.isclass(to_be_decorated):
-            logger.debug('The Decorator was applied to a class.')
-            return self.patched_importer_class(to_be_decorated)
-
-        elif inspect.isfunction(to_be_decorated):
-            logger.debug('The Decorator was applied to an instancemethod.')
-            return self.patched_extract_function(to_be_decorated)
-
-    def patched_importer_class(self, importer_class):
-        importer_class.extract = self.patched_extract_function(importer_class.extract)
-        return importer_class
-
-    def patched_extract_function(self, original_extract_function):
-        decorator = self
-
-        @wraps(original_extract_function)
-        def wrapper(self, file, existing_entries=None):
-
-            # read the importer's existing entries, if provided as argument to its `extract` method:
-            decorator.existing_entries = existing_entries
-
-            # read the importer's `extract`ed entries
-            logger.debug(f"About to call the importer's extract function to receive entries to be imported...")
-            if 'existing_entries' in inspect.signature(original_extract_function).parameters:
-                decorator.imported_entries = original_extract_function(self, file, existing_entries)
-            else:
-                decorator.imported_entries = original_extract_function(self, file)
-
-            # read the importer's file_account, to be used as default value for the decorator's known `account`:
-            if inspect.ismethod(self.file_account) and not decorator.account:
-                logger.debug("Trying to read the importer's file_account, "
-                             "to be used as default value for the decorator's `account` argument...")
-                file_account = self.file_account(file)
-                if file_account:
-                    decorator.account = file_account
-                    logger.debug(f"Read file_account {file_account} from the importer; "
-                                 f"using it as known account in the decorator.")
-                else:
-                    logger.debug(f"Could not retrieve file_account from the importer.")
-
-            return decorator.enhance_transactions()
-
-        return wrapper
-
-    def enhance_transactions(self):# load training data
+    def enhance_transactions(self):
+        # load training data
         self.training_data = ml.load_training_data(
             self.training_data,
             known_account=self.account,
